refactor(server): gallery scan consumes mold-db readers and walker
scan_gallery_dir shrinks to a ~25-line map over
mold_db::scan::scan_output_dir + metadata_io::read_or_synthesize,
deleting ~240 lines of duplicated validity guards and PNG/JPEG readers
(whose mold-db twins carried 'mirrors routes.rs' comments). Pure-fn
guard tests move to mold-db; the end-to-end scan tests stay in
routes.rs as the server regression net and now exercise the shared
stack. Server gallery fallback gains embedded-GIF metadata; the
now-unused png 0.18 dep is dropped from mold-server.
