+++ crates/mold-server/Cargo.toml
@@ -60,7 +60,6 @@ futures = "0.3"
 utoipa = { version = "5", features = ["preserve_order"] }
 sha2 = "0.10"
 walkdir = "2"
-png = "0.18"
 image = "0.25"
 governor = { version = "0.8", features = ["std"] }
 uuid = { version = "1", features = ["v4"] }
